add plot_lspeakperiod_vs_age
for blue stars, it gives like a factor of 2 in Prot over 2 dex in age.

uhhhh this is smaller than it should be, if Prot ~ t^{1/2}.

(that would imply a factor of 10 in Prot, over 2 dex in age).

But let's keep in mind the "convergence" only is supposed to happen at
like... 300 Myr or so, IIRC?

(different for different stellar masses).